requirements.txt
beautifulsoup4
requests
python-dotenv==1.0.1